sap xong rui

+++ src/page/cart-page.jsx
@@ -1,22 +1,41 @@
 import {
+  faCaretRight,
+  faCheck,
   faCircleMinus,
   faCirclePlus,
   faCircleXmark,
 } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { deleteDoc, doc } from "firebase/firestore";
+import {
+  collection,
+  deleteDoc,
+  doc,
+  query,
+  updateDoc,
+  where,
+} from "firebase/firestore";
 import { useContext, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
 import Footer from "~/components/footer";
 import Header from "~/components/header";
 import PageTitle from "~/components/page-title";
 import { firebase_store } from "~/config/firebase-config";
 import { AuthContext } from "~/context-api/auth-provider";
-
+import { BiUndo } from "react-icons/bi";
+import routes from "~/config/routes";
 const ShoppingCart = () => {
   const { cart } = useContext(AuthContext);
-  console.log(cart);
+  const formatNumber = (num) => {
+    return num.toLocaleString();
+  };
+  let result = 0;
+  cart.forEach((item) => {
+    if (item.price_sale) {
+      result += item.price_sale * item.quantity;
+    } else {
+      result += item.price * item.quantity;
+    }
+  });
   return (
     <>
       <Header />
@@ -37,6 +56,32 @@ const ShoppingCart = () => {
               {cart.map((product, index) => {
                 return <ProductItem product={product} key={index} />;
               })}
+            </div>
+            <div className="flex text-lg font-semibold mt-7">
+              <h2 className="pr-28">Tổng tiền</h2>{" "}
+              <p className="ml-28">{formatNumber(result)}đ</p>
+            </div>
+            <div className="flex">
+              <Link
+                to={routes.home}
+                className={`inline-flex items-center bg-sky-500 text-base mt-4 text-white px-8 h-12 rounded-3xl relative overflow-hidden group z-[1] mr-8 font-semibold`}
+              >
+                <span className="mr-1 font-bold text-2xl">
+                  <BiUndo />
+                </span>
+                Quay lại mua hàng
+                <div className="absolute top-0 left-0 right-0 bottom-0 bg-sky-400 w-0 transition-all ease-linear duration-200 z-[-1] group-hover:w-full"></div>
+              </Link>
+              <Link
+                to={routes.payPage}
+                className={`inline-flex items-center bg-sky-500 text-base mt-4 text-white px-8 h-12 rounded-3xl relative overflow-hidden group z-[1] mr-8 font-semibold`}
+              >
+                <span className="mr-1 font-bold text-xl">
+                  <FontAwesomeIcon icon={faCheck} />
+                </span>
+                Tiến hành thanh toán
+                <div className="absolute top-0 left-0 right-0 bottom-0 bg-sky-400 w-0 transition-all ease-linear duration-200 z-[-1] group-hover:w-full"></div>
+              </Link>
             </div>
           </>
         )}
@@ -59,9 +104,33 @@ const ProductItem = ({ product }) => {
   const formatNumber = (num) => {
     return num.toLocaleString();
   };
+  const { user } = useContext(AuthContext);
 
   const deleteProduct = async () => {
-    await deleteDoc(doc(firebase_store, "cart", `product${product.id}`));
+    await deleteDoc(doc(firebase_store, "cart", `product${product.id}_${user.uid}`));
+  };
+
+  const UpdateProduct = async (type) => {
+    const washingtonRef = doc(
+      firebase_store,
+      "cart",
+      `product${product.id}_${user.uid}`
+    );
+
+    // Set the "capital" field of the city 'DC'
+    if (type === "PUSH") {
+      await updateDoc(washingtonRef, {
+        quantity: product.quantity + 1,
+      });
+    } else {
+      if (product.quantity == 1) {
+        deleteProduct();
+      } else {
+        await updateDoc(washingtonRef, {
+          quantity: product.quantity - 1,
+        });
+      }
+    }
   };
 
   return (
@@ -94,20 +163,16 @@ const ProductItem = ({ product }) => {
         )}
       </div>
       <div className="text-[#444] bg-[#ebebeb] py-2 justify-between flex w-32 items-center rounded-3xl overflow-hidden">
-        <span className="py-1 px-2 cursor-pointer">
+        <span
+          className="py-1 px-2 cursor-pointer"
+          onClick={() => UpdateProduct("REMOVE")}
+        >
           <FontAwesomeIcon icon={faCircleMinus} />
         </span>
         <span> {product.quantity} </span>
         <span
           className="py-1 px-2 cursor-pointer"
-          onClick={(e) => {
-            // dispatch(
-            //   add_by_cart({
-            //     ...product,
-            //     quantity: product.quantity + 1,
-            //   })
-            // );
-          }}
+          onClick={() => UpdateProduct("PUSH")}
         >
           <FontAwesomeIcon icon={faCirclePlus} />
         </span>
